requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
google-generativeai==0.8.3
pydantic==2.9.2
python-dotenv==1.0.1
pymongo==4.8.0
certifi>=2024.2.2